requirements.txt
# YouTube Data API
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0

# Data Processing
pandas==2.0.3
numpy==1.24.3
scipy==1.11.2

# Natural Language Processing
nltk==3.8.1
textblob==0.17.1
transformers==4.35.0
torch==2.0.1

# Spark and Big Data Processing
pyspark==3.5.0

# Sentiment Analysis
vader-lexicon==2.1.0
scikit-learn==1.3.1

# Visualization & Dashboard
streamlit==1.28.0
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2

# Task Scheduling
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
joblib==1.3.2
python-dateutil==2.8.2

# Development
jupyter==1.0.0
ipython==8.15.0
pytest==7.4.2